// src/components/organisms/CreateCompanyForm/types.ts
export interface handleColorProps {
  color: any;
  switchValue: number;
}

export interface addressStateProps {
  street: string;
  number: number;
  complement?: string;
  neighborhood: string;
  city: string;
  state: string;
}

export interface colorStateProps {
  primaryColor: string;
  secondaryColor: string;
}